backend/requirements.txt
django~=3.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
Pillow==9.1.1
sorl-thumbnail==12.8.0
djangorestframework==3.13.1
